requirements.txt
absl-py==2.1.0
accelerate==0.25.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
ansi2html==1.9.1
anyconfig==0.14.0
anyio==4.2.0
appdirs==1.4.4
arrow==1.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
blessed==1.20.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constants==0.6.0
contourpy==1.2.0
controlnet-aux==0.0.7
crc32c==2.3.post0
cycler==0.12.1
datasets==2.16.1
deepspeed==0.13.1
diffusers==0.25.1
dill==0.3.7
dm-tree==0.1.8
docker-pycreds==0.4.0
easydict==1.11
ecdsa==0.18.0
einops==0.7.0
exceptiongroup==1.2.0
fairscale==0.4.13
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.13.1
Flask==2.1.3
Flask-RESTful==0.3.10
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gpustat==1.1.1
gradio==4.16.0
gradio_client==0.8.1
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
hjson==3.1.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
imageio==2.33.1
imgaug==0.4.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.0.4
kiwisolver==1.4.5
lazy_loader==0.3
lightning==2.1.3
lightning-utilities==0.10.1
ligo-segments==1.4.0
lmdb==1.4.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
namex==0.0.7
natsort==8.4.0
networkx==3.2.1
ninja==1.11.1.1
nose==1.3.7
numpy==1.26.3
nvgpu==0.10.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.550.52
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.9.0.80
orjson==3.9.12
packaging==23.2
pandas==2.2.0
peft==0.7.1
pillow==10.2.0
prodigyopt==1.0
protobuf==3.20.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-Levenshtein==0.12.2
python-multipart==0.0.6
pytorch-lightning==2.1.3
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
ruff==0.1.14
sacremoses==0.1.1
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.12.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.39.2
seqeval==0.0.12
setproctitle==1.3.3
shapely==2.0.2
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.35.1
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
tifffile==2023.12.9
timm==0.9.12
tokenizers==0.13.3
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.2
torchmetrics==1.3.0.post0
torchvision==0.16.2
tqdm==4.66.1
transformers==4.31.0
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.27.0
wandb==0.16.2
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==2.0.3
xformers==0.0.23.post1
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
